backend/requirements.txt
asgiref==3.5.2
Django==4.1.1
django-cors-headers==3.13.0
django-environ==0.9.0
gunicorn==20.1.0
sqlparse==0.4.2
